split lift
