+++ src/rpp/rpp/operators/fwd.hpp
@@ -21,6 +21,7 @@
 #include <rpp/operators/fwd/flat_map.hpp>
 #include <rpp/operators/fwd/group_by.hpp>
 #include <rpp/operators/fwd/last.hpp>
+#include <rpp/operators/fwd/lift.hpp>
 #include <rpp/operators/fwd/map.hpp>
 #include <rpp/operators/fwd/merge.hpp>
 #include <rpp/operators/fwd/multicast.hpp>
